+++ index.js
@@ -188,6 +188,10 @@ var _isCreditCard = require('./lib/isCreditCard');
 
 var _isCreditCard2 = _interopRequireDefault(_isCreditCard);
 
+var _isIdentityCard = require('./lib/isIdentityCard');
+
+var _isIdentityCard2 = _interopRequireDefault(_isIdentityCard);
+
 var _isISIN = require('./lib/isISIN');
 
 var _isISIN2 = _interopRequireDefault(_isISIN);
@@ -347,6 +351,7 @@ var validator = {
   isBefore: _isBefore2.default,
   isIn: _isIn2.default,
   isCreditCard: _isCreditCard2.default,
+  isIdentityCard: _isIdentityCard2.default,
   isISIN: _isISIN2.default,
   isISBN: _isISBN2.default,
   isISSN: _isISSN2.default,